fix: infer optional arguments in SDK methods
This change enables a flag that will cause the generator to infer when request and security arguments of SDK methods should be treated as optional. The rule is: if all fields of a request or security type are optional then an argument of that type will be marked as optional in methods.